tests/wpt/web-platform-tests/mediacapture-streams/MediaDevices-SecureContext.html

<!doctype html>
<html>
<head>
<title>MediaDevices and SecureContext</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
</head>
<body>
<script>

assert_false(window.isSecureContext, "This test must be run in a non secure context");
assert_false('MediaDevices' in window, "MediaDevices is not exposed");
assert_false('getUserMedia' in navigator, "getUserMedia is not exposed");
assert_false('mediaDevices' in navigator, "mediaDevices is not exposed");

done();
</script>
</body>
</html>
